[INTERPRETER] Use acc's dtype as computation dtype in `tl.dot` (#3566)
I found with `TRITON_INTERPRET=1`, `tl.dot` almost always gives wrong
result for INT8 input, due to `np.matmul` use inputs' dtype as result
dtype by default:

```
>>> a = np.array([[10, 20]], np.int8)
>>> np.matmul(a, a.transpose())
array([[-12]], dtype=int8)
>>> np.matmul(a, a.transpose(), dtype=np.int32)
array([[500]])
```

+++ python/triton/runtime/interpreter.py
@@ -411,7 +411,7 @@ class InterpreterBuilder:
         return TensorHandle(np.transpose(arg.data, perm), arg.dtype)
 
     def create_dot(self, a, b, d, input_precision, max_num_imprecise_acc):
-        return TensorHandle(np.matmul(a.data, b.data) + d.data, d.dtype)
+        return TensorHandle(np.matmul(a.data, b.data, dtype=d.data.dtype) + d.data, d.dtype)
 
     def create_make_range(self, start, stop):
         return TensorHandle(np.arange(start, stop, dtype=np.int32), tl.int32)
